podagent/requirements.txt
fastapi
uvicorn
tqdm
sentence-transformers
faiss-cpu
transformers
torch
rouge-score
bert-score
pydantic
pypdf
openai
together
